Reapply "update_attacking"
This reverts commit 2fbe9bd31509c1f34d02e1d9c28f5b2ae52ac89c.

+++ StickDefence/Assets/Scripts/Models/SO/Core/StickmanStatsConfig.cs
@@ -9,7 +9,7 @@ namespace Models.SO.Core
         public PlayerUnitTypeEnum UnitType;
         public int Level;
         public int Damage;
-        public float Reloading;
+        public int Reloading;
         public int Price;
         public float AttackSpeed;
